test: 전역 상태 관리 테스트 (#29)
